first design for external controller
- still missing effects such as death and division

+++ src/backend/cpu_os_threads/domain_decomposition.rs
@@ -563,7 +563,7 @@ pub struct MultiVoxelContainer<
         >,
     >,
 
-    pub mvc_id: u16,
+    pub mvc_id: u32,
 }
 
 impl<
